+++ clients/drcachesim/simulator/caching_device.h
@@ -68,8 +68,9 @@ class caching_device_t {
 public:
     explicit caching_device_t(const std::string &name = "caching_device");
     virtual bool
-    init(int associativity, int64_t block_size, int num_blocks, caching_device_t *parent,
-         caching_device_stats_t *stats, prefetcher_t *prefetcher = nullptr,
+    init(int associativity, int64_t block_size, int64_t num_blocks,
+         caching_device_t *parent, caching_device_stats_t *stats,
+         prefetcher_t *prefetcher = nullptr,
          cache_inclusion_policy_t inclusion_policy =
              cache_inclusion_policy_t::NON_INC_NON_EXC,
          bool coherent_cache = false, int id_ = -1,
@@ -148,7 +149,7 @@ public:
     {
         return block_size_;
     }
-    virtual int
+    virtual int64_t
     get_num_blocks() const
     {
         return num_blocks_;
@@ -205,7 +206,7 @@ protected:
     {
         return addr >> block_size_bits_;
     }
-    inline int
+    inline int64_t
     compute_block_idx(addr_t tag) const
     {
         return (tag & blocks_per_way_mask_) * associativity_;
@@ -246,7 +247,7 @@ protected:
 
     int associativity_;
     int64_t block_size_; // Also known as line length.
-    int num_blocks_;     // Total number of lines in cache = size / block_size.
+    int64_t num_blocks_; // Total number of lines in cache = size / block_size.
     bool coherent_cache_;
     // This is an index into snoop filter's array of caches.
     int id_;
@@ -267,9 +268,9 @@ protected:
     // an extended block class which has its own member variables cannot be indexed
     // correctly by base class pointers.
     caching_device_block_t **blocks_;
-    int blocks_per_way_;
+    int64_t blocks_per_way_;
     // Optimization fields for fast bit operations
-    int blocks_per_way_mask_;
+    int64_t blocks_per_way_mask_;
     int block_size_bits_;
 
     caching_device_stats_t *stats_;
